Message

+++ Desktop/Project/shopify/src/app/admin/admin-user/admin-user.component.ts
@@ -29,7 +29,7 @@ export class AdminUserComponent {
 
   users: any[] = []
 
-  ngOnInit() {
+  getUsers() {
     forkJoin([
       this.http.get<any[]>('http://localhost:3000/users'),
       this.http.get<any[]>('http://localhost:3000/orders')
@@ -40,6 +40,10 @@ export class AdminUserComponent {
         return { ...user, totalProducts };
       });
     });
+  }
+
+  ngOnInit() {
+    this.getUsers()
   }
 
   calculate(dateOfBirth: any) {  
@@ -74,13 +78,23 @@ export class AdminUserComponent {
   showDialog(user: any) {
     this.isEditMode = true;
     this.adduser = true;
+  
     this.id = user.id;
     this.username = user.username;
     this.firstname = user.firstname;
     this.lastname = user.lastname;
     this.gender = user.gender;
     this.email = user.email;
-    this.phone = user.phone;
+  
+    const phoneMatch = user.phone?.match(/^(\+\d{1,3})(\d{10})$/);
+    if (phoneMatch) {
+      this.selectedCountryCode = phoneMatch[1];
+      this.phone = phoneMatch[2];
+    } else {
+      this.selectedCountryCode = '';
+      this.phone = '';
+    }
+  
     this.address = user.address;
     this.country = user.country;
     this.state = user.state;
@@ -89,8 +103,22 @@ export class AdminUserComponent {
     this.locale = user.locale;
     this.admin = user.admin;
     this.imageUrl = user.imageUrl;
-    this.filteredStates = this.states[this.country.code] || [];
+  
+    if (user.dob) {
+      const [day, month, year] = user.dob.split('/').map(Number);
+      this.dob = new Date(year, month - 1, day);
+    } else {
+      this.dob = null;
+    }
+
+    if (this.country?.code) {
+      this.filteredStates = this.states[this.country.code] || [];
+    } else {
+      this.filteredStates = [];
+    }
   }
+  
+  
   
   DeleteShow: boolean = false
   showDelete(user: any) {
@@ -98,7 +126,6 @@ export class AdminUserComponent {
     this.DeleteShow = true;
   }
   
-
     cancelButton() {
       this.DeleteShow = false
     }
@@ -123,14 +150,19 @@ export class AdminUserComponent {
       this.conformpass = '';
       this.filteredStates = [];
       this.adduser = true;
+      this.resetForm();
     }
 
 
-    confirmDelete(user: any) {
-      
+    confirmDelete() {
+      let user = this.selectedUser
+      if (this.loggedUser.username === user.username) {
+        this.messageService.add({ severity: 'error', summary: 'Your not able delete', detail: "It's Your Profile" });
+        return
+      }
       this.http.get<any[]>(`http://localhost:3000/orders?username=${user.username}`).subscribe({
         next: (orders) => {
-          
+          console.log(orders)
           const deleteOrderRequests = orders.map(order =>
             this.http.delete(`http://localhost:3000/orders/${order.id}`)
           );
@@ -168,9 +200,9 @@ export class AdminUserComponent {
         }
       });
     }
-
-    
+   
     cancelUser() {
+      this.resetForm();
       this.adduser = false
     }
 
@@ -178,7 +210,8 @@ export class AdminUserComponent {
 
     //form Data
     messageService: MessageService = inject(MessageService)
-      datacollection: FormdatacollectionService = inject(FormdatacollectionService)
+    datacollection: FormdatacollectionService = inject(FormdatacollectionService)
+    @ViewChild('formData') data: NgForm
     
        username: string = "";
        show: boolean = true;
@@ -216,7 +249,12 @@ export class AdminUserComponent {
         'en-IN', 'en-US', 'fr-FR', 'de-DE'
       ];
 
-      countryCode = ['+91', '+01']
+      countryCodeList = [
+        { label: '+91', value: '+91' },
+        { label: '+01', value: '+01' }
+      ];
+      
+      selectedCountryCode: string = '';
     
       country: any = '';
       filteredStates: any[] = [];
@@ -230,7 +268,6 @@ export class AdminUserComponent {
         this.state = null;
       }
     
-      @ViewChild('formData') data: NgForm
     
       onChangeImage(event: any) {
         const file = event.target.files[0];
@@ -254,14 +291,17 @@ export class AdminUserComponent {
         }
       }
       
-    
+      getInitials(): string {
+        const initials = `${this.firstname.charAt(0)}${this.lastname.charAt(0)}`;
+        return `https://ui-avatars.com/api/?name=${initials}&background=random`;
+      }
       
       submitData() {
-        if (this.data.invalid) {
+        if (this.data.invalid || !this.selectedCountryCode || !this.country || !this.state) {
           this.messageService.add({
             severity: 'error',
             summary: 'Form Invalid',
-            detail: 'Please fill all required fields correctly.'
+            detail: 'Please fill all required fields.'
           });
           return;
         }
@@ -269,20 +309,20 @@ export class AdminUserComponent {
         if (this.password !== this.conformpass) {
           this.messageService.add({
             severity: 'error',
-            summary: 'Password Mismatched',
+            summary: 'Password Mismatch',
             detail: 'Passwords do not match!'
           });
           return;
         }
       
         const formData = {
-          id: this.id,
+          username: this.username,
           firstname: this.firstname,
           lastname: this.lastname,
           gender: this.gender,
           dob: this.formatDate(this.dob),
           email: this.email,
-          phone: this.phone,
+          phone: `${this.selectedCountryCode}${this.phone}`,
           address: this.address,
           country: this.country,
           state: this.state,
@@ -292,76 +332,119 @@ export class AdminUserComponent {
           admin: this.admin,
           password: this.password,
           conformpass: this.conformpass,
-          imageUrl: this.imageUrl
+          imageUrl: this.imageUrl === "../../assets/profile.png" ? this.getInitials() : this.imageUrl
         };
       
-        // EDIT MODE
-        if (this.isEditMode) {
-          const index = this.users.findIndex(u => u.id === this.id);
-          if (index !== -1) {
-            this.users[index] = {
-              ...this.users[index],
-              firstname: this.firstname,
-              lastname: this.lastname,
-              gender: this.gender,
-              dob: this.formatDate(this.dob),
-              email: this.email,
-              phone: this.phone,
-              address: this.address,
-              country: this.country,
-              state: this.state,
-              Zipcode: this.zipCode,
-              timeZone: this.timeZone,
-              locale: this.locale,
-              imageUrl: this.imageUrl
-            };
+        forkJoin({
+          emailExists: this.datacollection.isEmailDuplicate(this.email),
+          usernameExists: this.datacollection.isDuplicate(this.username)
+        }).subscribe({
+          next: ({ emailExists, usernameExists }) => {
+            if (this.isEditMode) {
+              const currentUser = this.users.find(u => u.id === this.id);
       
-            this.http.put(`http://localhost:3000/users/${this.id}`, formData).subscribe({
+              const isDuplicateEmail = emailExists && this.email !== currentUser.email;
+              const isDuplicateUsername = usernameExists && this.username !== currentUser.username;
+      
+              if (isDuplicateEmail) {
+                this.messageService.add({
+                  severity: 'error',
+                  summary: 'Duplicate Email',
+                  detail: 'Email already exists!'
+                });
+                return;
+              }
+      
+              if (isDuplicateUsername) {
+                this.messageService.add({
+                  severity: 'error',
+                  summary: 'Duplicate Username',
+                  detail: 'Username already exists!'
+                });
+                return;
+              }
+      
+              const index = this.users.findIndex(u => u.id === this.id);
+              if (index !== -1) {
+                this.users[index] = { ...this.users[index], ...formData };
+              }
+      
+              this.http.put(`http://localhost:3000/users/${this.id}`, formData).subscribe({
+                next: () => {
+                  this.messageService.add({
+                    severity: 'success',
+                    summary: 'Updated',
+                    detail: 'User updated successfully!'
+                  });
+                  this.resetForm();
+                  this.adduser = false;
+                  this.isEditMode = false;
+
+                },
+                error: () => {
+                  this.messageService.add({
+                    severity: 'error',
+                    summary: 'Error',
+                    detail: 'Failed to update user.'
+                  });
+                }
+              });
+              return;
+            }
+      
+            if (emailExists) {
+              this.messageService.add({
+                severity: 'error',
+                summary: 'Duplicate Email',
+                detail: 'Email already exists!'
+              });
+              return;
+            }
+      
+            if (usernameExists) {
+              this.messageService.add({
+                severity: 'error',
+                summary: 'Duplicate Username',
+                detail: 'Username already exists!'
+              });
+              return;
+            }
+      
+            this.datacollection.postData(formData).subscribe({
               next: () => {
                 this.messageService.add({
                   severity: 'success',
-                  summary: 'Updated',
-                  detail: 'User updated successfully!'
+                  summary: 'Success',
+                  detail: 'User added successfully'
                 });
+                this.getUsers()
+                this.resetForm();
                 this.adduser = false;
-                this.isEditMode = false;
               },
               error: () => {
                 this.messageService.add({
                   severity: 'error',
                   summary: 'Error',
-                  detail: 'Failed to update user.'
+                  detail: 'Form submission failed'
                 });
               }
             });
-          }
-          return;
-        }
-      
-        // ADD MODE
-
-        this.datacollection.postData(formData).subscribe({
-          next: () => {
-            this.messageService.add({
-              severity: 'success',
-              summary: 'Success',
-              detail: 'User added successfully'
-            });
-            this.adduser = false;
-            this.route.navigate(['/login']);
           },
           error: () => {
             this.messageService.add({
               severity: 'error',
               summary: 'Error',
-              detail: 'Form submission failed'
+              detail: 'Validation failed!'
             });
           }
         });
       }
       
+      
+      
       cancelBtn() {
         this.adduser = false
+        this.resetForm();
       }
 
 
@@ -394,4 +477,29 @@ export class AdminUserComponent {
     deleteProfile() {
       this.imageUrl = '../../assets/profile.png'
     }
+
+    resetForm() {
+      this.data?.resetForm();
+      this.username = '';
+      this.firstname = '';
+      this.lastname = '';
+      this.gender = 'Male';
+      this.email = '';
+      this.phone = '';
+      this.selectedCountryCode = '';
+      this.address = '';
+      this.country = '';
+      this.state = '';
+      this.zipCode = '';
+      this.timeZone = '';
+      this.locale = '';
+      this.password = '';
+      this.conformpass = '';
+      this.admin = false;
+      this.imageUrl = '../../assets/profile.png';
+      this.selectedFile = null;
+      this.dob = undefined;
+      this.filteredStates = [];
+    }
+    
 }
